Really cool git resource
I want to recommend this free Git/GitHub course on Udacity.  I started this wayyyy back in May and forgot about it, got back round to it a couple of weeks ago and found it **super, super** useful, literally learnt so much.  It explains the differences between Git and GitHub clearly, using the staging area properly, using `git diff` properly to find your bugs.  And it explains it **clearly**.  Plus it teaches you by getting you to try Git for yourself on your machine.  

(Basically it's amazing.)

I feel like doing this before starting the course will possibly eliminate the *OMG-WHAT-IS-THIS-GIT-THING* rite of passage for new students :wink: